feat(fsd): A2 E2 — receivables + dunning を entities へ (#376)
型のみ移設（Strategy R・挙動保存）:
- entities/manual-receivable (ManualReceivable + ManualReceivableImportResult・ADR 0014 コメント同梱)
- entities/dunning-notice    (DunningNotice)

consumer 再ポイント: endpoints.ts の型 block・manual-receivables page + test。
DunningNotice の外部 consumer は endpoints.ts のみ。DunningPause/Preview/Stage
等の request/response 型は endpoints.ts に残置（A1 が entities/*/api-types.ts へ）。

logic-diff-0 証明: types/index.ts は 3 interface の削除のみ（+0 行）・移設 body
md5 一致。新規行は header doc comment のみ。npm run check 緑（vitest 62 不変）。

Refs #376
