+++ frontend/src/entities/dunning-notice/index.ts
@@ -1,0 +1,1 @@
+export type { DunningNotice } from './model'
