+++ frontend/src/entities/manual-receivable/index.ts
@@ -1,0 +1,1 @@
+export type { ManualReceivable, ManualReceivableImportResult } from './model'
